pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentrace"
version = "1.0.0"
description = "Framework-agnostic evaluation and tracing for LLM agentic pipelines (AgentRace / AgentTrace)."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{ name = "AgentRace contributors" }]
keywords = ["llm", "agents", "evaluation", "tracing", "observability"]
dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic-settings>=2.0.0",
    "aiosqlite>=0.19.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
postgres = [
    "asyncpg>=0.29.0",
]
judge = [
    "litellm>=1.40.0",
]
all = [
    "asyncpg>=0.29.0",
    "litellm>=1.40.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.0.0",
    "mypy>=1.8.0",
    "ruff>=0.15.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
agentrace = "cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentrace*", "cli*"]

[tool.setuptools.package-data]
agentrace = ["py.typed"]
"agentrace.dataset.benchmarks" = ["*.json"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = [
    "E501",  # line length (handled by formatter where practical)
]

[tool.ruff.lint.per-file-ignores]
# Imports after conditional ``evaluate`` binding and SDK patch side effects.
"agentrace/__init__.py" = ["E402"]

# Optional capture adapters import third-party SDKs that are not core deps.
[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = [
    "llama_index.*",
    "langchain_core.*",
    "crewai",
    "anthropic.*",
]
ignore_missing_imports = true
